+++ python/samples/04.ai.g.autogen.product-spec-critique/PM-spec-barebones.md
@@ -1,0 +1,4 @@
+# Product Management Specification: Automatic Bubble Blower
+
+## Product Overview
+The Automatic Bubble Blower is a fun and interactive device designed to keep children entertained by continuously producing bubbles. Its user-friendly design ensures ease of use for parents while providing endless amusement for kids. The device is portable, safe, and equipped with advanced features to enhance the user experience.
